wlfreerdp: initial Wayland client
Implement an initial Wayland client, which will build if
the wayland-client development libraries are detected
(or if -DWITH_WAYLAND:BOOL=ON is set). It is currently
view-only, but inputs will be implemented soon.

It uses the software SHM interface, which means it does not
require GL acceleration to run. It should be compatible
with any compositor.

Change-Id: I00302327f167f1ebc20b65aacb08bd65cf6e8487
Origin: https://github.com/FreeRDP/FreeRDP/pull/2145
Signed-off-by: Manuel Bachmann <tarnyko@tarnyko.net>